requirements.txt
Cerberus==1.1
Eve==0.8
Flask==0.12
Flask-PyMongo==0.5.1
pymongo==3.5.1
requests==2.18.4
requests-toolbelt==0.8.0
urllib3==1.22
Werkzeug==0.11.15
python-telegram-bot==8.1.1